# backend/static/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Store App API",
    "description": "An API for store app"
  },
  "paths": {
    "/categories": {
      "get": {
        "summary": "Get categories list",
        "tags": [
          "Category"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "summary": "Create new category",
        "tags": [
          "Category"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Category object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "/types": {
      "get": {
        "summary": "Get types list",
        "tags": [
          "Type"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "summary": "Create new type",
        "tags": [
          "Type"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Type object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Type"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "/products": {
      "get": {
        "summary": "Get products list",
        "tags": [
          "Product"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "summary": "Create new product",
        "tags": [
          "Product"
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Product object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Category": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "Type": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "Product": {
      "type": "object",
      "required": [
        "name",
        "sku",
        "category_id",
        "type_id"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "sku": {
          "type": "string"
        },
        "stocks": {
          "type": "integer",
          "format": "int64",
          "minimun": 0
        },
        "stock_reserve": {
          "type": "integer",
          "format": "int64",
          "required": false,
          "minimun": 1
        },
        "category_id": {
          "type": "integer",
          "format": "int64"
        },
        "type_id": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
...
  }
}